+++ src/core/model/string_pool.cpp
@@ -1,20 +1,20 @@
-#include "scav/scav_string_pool.h"
+#include "scav/scav_core.h"
 
 #include "core/model/bytes_view.h"
-#include "scav/scav_ids.h"
 #include "scav/scav_types.h"
 
+#include <cstddef>
 #include <cstdint>
 #include <string_view>
 
 namespace scav {
 
 int string_compare_bytes(scav_byte const *a,
-                         uint32_t alen,
+                         size_t alen,
                          scav_byte const *b,
-                         uint32_t blen) {
-  uint32_t const n{ (alen < blen) ? alen : blen };
-  for (uint32_t i = 0; i < n; ++i) {
+                         size_t blen) {
+  size_t const n{ (alen < blen) ? alen : blen };
+  for (size_t i = 0; i < n; ++i) {
     if (a[i] != b[i]) { return (a[i] < b[i]) ? -1 : 1; }
   }
   if (alen == blen) { return 0; }
